extract path traversal
